extended functionality
Extended functionality to pull description. Description text has extra spaces removed and any double quote changed to single quote. Recommend that the scripts  to scrape data and load data be separate so that any problems in the data can be looked at and cleaned before loading it into the database. Especially as we don't really need to run the extractor more than about once anyway.